Converted query prepare to functional
